Split provider packaging from release finalization
Separate provider archive creation from release metadata finalization so CI can build platform archives independently and then finalize a merged dist directory once. The package command now invalidates stale finalization files while preserving archives for other platform targets in the same dist.
